Update README to point to new repository
Update the README to point users to the new repository, including a
short migration guide.

We shouldn't land this on a Friday :) Ideally, we'd first redirect all traffic
from https://connect.build to https://connectrpc.com: this clearly indicates
that we're moving, and IMO minimizes confusion. Once all web traffic is going
to the new site, we can merge all these deprecation PRs.